Add new_simple_exception; fill SyntaxError and f_lineno
new_exception constructs extra-payload types; BaseException-sized
types use new_simple_exception. Non-UTF-8 SyntaxError sets location
and text. Codegen errors load text via ProgramText. Keep Missing
parentheses capitalization. f-string adjacent atoms report a missing
comma. f_lineno may be None; lasti is not restored on unwind.

+++ crates/vm/src/stdlib/_testinternalcapi.rs
@@ -368,7 +368,7 @@ mod _testinternalcapi {
     }
 
     #[pyfunction]
-    fn iframe_getline(frame: PyRef<FrameObject>) -> usize {
+    fn iframe_getline(frame: PyRef<FrameObject>) -> i32 {
         frame.f_lineno()
     }
